JCLOUDS-703 missing imageSpaceGb from machineType
The machine types don't seem to be giving a value for imageSpaceGb.
This means they get the value zero and as a consequence the Hardware does
not support an image.

However, google's documentation states that 'You must always specify an
image when you create an instance...'.

So it would seem sensible that the missing value for this is ignored and
supportsImage uses the alwaysTrue predicate.